backend/api/v1/endpoints/media.py
from fastapi import APIRouter, HTTPException, UploadFile, File, Form
from typing import Optional
import io
from PIL import Image
import numpy as np

from services.mie_service import MIEService

router = APIRouter()

@router.post("/verify")
async def verify_media(
    file: UploadFile = File(...),
    media_type: str = Form(...)
):
    """
    Verify media authenticity using AI Media Integrity Engine.
    
    Args:
        file: Image or video file to analyze
        media_type: "image" or "video"
    
    Returns:
        Trust score (0-100) and detailed analysis
    """
    try:
        # Validate file type
        if media_type not in ["image", "video"]:
            raise HTTPException(status_code=400, detail="media_type must be 'image' or 'video'")
        
        # Read file content
        file_content = await file.read()
        
        # Initialize MIE service
        mie_service = MIEService()
        
        if media_type == "image":
            result = await mie_service.analyze_image(file_content, file.filename)
        else:
            result = await mie_service.analyze_video(file_content, file.filename)
        
        return {
            "trust_score": result["trust_score"],
            "confidence": result["confidence"],
            "analysis_details": {
                "face_detection": result.get("face_detection", {}),
                "temporal_consistency": result.get("temporal_consistency", {}),
                "visual_artifacts": result.get("visual_artifacts", {}),
                "metadata_analysis": result.get("metadata_analysis", {})
            },
            "recommendations": result.get("recommendations", []),
            "processing_time": result.get("processing_time", 0)
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error verifying media: {str(e)}")

@router.get("/verify/{verification_id}")
async def get_verification_result(verification_id: str):
    """
    Get cached verification result by ID.
    
    For cases where verification takes time and is processed asynchronously.
    """
    try:
        # TODO: Implement Redis cache lookup
        # For now, return placeholder
        return {
            "verification_id": verification_id,
            "status": "completed",
            "message": "Verification cache lookup not yet implemented"
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error retrieving verification: {str(e)}")

@router.post("/batch_verify")
async def batch_verify_media(
    files: list[UploadFile] = File(...)
):
    """
    Batch verify multiple media files.
    
    Returns array of verification results.
    """
    try:
        results = []
        mie_service = MIEService()
        
        for file in files:
            try:
                file_content = await file.read()
                
                # Determine media type from file extension
                file_ext = file.filename.lower().split('.')[-1] if file.filename else ""
                if file_ext in ['jpg', 'jpeg', 'png', 'bmp', 'gif']:
                    media_type = "image"
                elif file_ext in ['mp4', 'avi', 'mov', 'webm', 'mkv']:
                    media_type = "video"
                else:
                    results.append({
                        "filename": file.filename,
                        "error": "Unsupported file type"
                    })
                    continue
                
                if media_type == "image":
                    result = await mie_service.analyze_image(file_content, file.filename)
                else:
                    result = await mie_service.analyze_video(file_content, file.filename)
                
                results.append({
                    "filename": file.filename,
                    "trust_score": result["trust_score"],
                    "confidence": result["confidence"],
                    "media_type": media_type
                })
                
            except Exception as e:
                results.append({
                    "filename": file.filename,
                    "error": str(e)
                })
        
        return {"results": results}
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error in batch verification: {str(e)}")

@router.get("/stats")
async def get_verification_stats():
    """
    Get verification statistics and model performance metrics.
    """
    try:
        # TODO: Implement real statistics from database
        return {
            "total_verifications": 0,
            "average_trust_score": 0.0,
            "model_accuracy": 0.92,
            "processing_time_avg": 2.3,
            "supported_formats": {
                "images": [".jpg", ".jpeg", ".png", ".bmp"],
                "videos": [".mp4", ".avi", ".mov", ".webm"]
            }
        }
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Error retrieving stats: {str(e)}")